Install react app, delete src folder, add two files
After installing react by typing in 'npx create-react-app ./' the '/ installs react in the current folder. Next thing that was done was that I deleted the src folder with everything in it and then created a new one and created a 'index.js' file and a 'App.jsx' file' The index.js file has the 3 imports: React, ReactDOM and App. Finally I add the ReactDOM.render....