Add to App.jsx file, installed materialui, copy and past to public > index.html
In App.js I start by importing React from react then I create an App function and set it as a const type which is similar to making an object in Java final. In the bracket of the App function/method there is a return function and in its parameter it expresses an html Hello world. The next thing I did was installed material ui by typing 'npm install @material-ui/core'. The third thing I did was go to the site and under 'Load the default Roboto font' I copied and pasted and html/style code in the index.html file above <title> in the public folder